# pool the transient maps used for Msg Pack, Truncate and Len