--- a/msg_truncate.go
+++ b/msg_truncate.go
@@ -54,7 +54,7 @@ func (dns *Msg) Truncate(size int) {
 		size -= Len(edns0)
 	}
 
-	compression := make(map[string]struct{})
+	compression := compressionPool.Get().(map[string]struct{})
 
 	l = headerSize
 	for _, r := range dns.Question {
@@ -88,6 +88,11 @@ func (dns *Msg) Truncate(size int) {
 		// Add the OPT record back onto the additional section.
 		dns.Extra = append(dns.Extra, edns0)
 	}
+
+	for k := range compression {
+		delete(compression, k)
+	}
+	compressionPool.Put(compression)
 }
 
 func truncateLoop(rrs []RR, size, l int, compression map[string]struct{}) (int, int) {