@miekg 2019-09-21
thanks!

this is somewhat concerning:
~~~
BenchmarkMuxMatch/uppercase-12                     100           105           +5.00%
BenchmarkServe-12                                  130004        141492        +8.84%
BenchmarkServe6-12                                 143604        137140        -4.50%
~~~
As this does include some (localhost) networking it might not be conclusive, is `Server-12` really almost 10% slower?