@tmthrgd 2019-09-21
I don’t believe this is a correct usage of sync.Pool, see golang/go#23199.