@charlievieth 2019-09-21
> As this does include some (localhost) networking it might not be conclusive, is `Server-12` really almost 10% slower?

There is a ton of variability in timing of the `Serve` benchmarks (MBP 2018, macOS 10.14.6, Core i9 @2.9 GHz).  TBH, they're only good for memory/alloc stats.  

Results from [`benchstat`](https://godoc.org/golang.org/x/perf/cmd/benchstat) (thanks @tmthrgd for showing me this tool) 
 below:
```
name              time/op
Serve-12           234µs ±59%
Serve6-12          231µs ±49%
ServeCompress-12   219µs ±68%

name              alloc/op
Serve-12          3.27kB ± 0%
Serve6-12         3.11kB ± 0%
ServeCompress-12  3.27kB ± 0%

name              allocs/op
Serve-12            53.0 ± 0%
Serve6-12           50.0 ± 0%
ServeCompress-12    53.0 ± 0%
```