@charlievieth 2019-09-21
> I don’t believe this is a correct usage of sync.Pool, see [golang/go#23199](https://github.com/golang/go/issues/23199).

I think the specific issue that you're referring to is that we are not setting an upper bound on the size of objects that we store in the pool.

I'm not terribly familiar with this part of DNS, but maybe @miekg could weigh in on what an appropriate upper limit would be (for reference `BenchmarkPackMsgMassive` stores 164 elements in the map)?