@miekg 2019-09-22
This does look to have real world impact... ?
~~~
name              time/op
Serve-12           234µs ±59%
Serve6-12          231µs ±49%
ServeCompress-12   219µs ±68%
~~~

We storing compression pointers in the Pool; there is an upper limit to this (before we know we need them), so every  different label (list) in a reply is stored. Its kinda hard to say what the exact limit is, but lets say:

* max size 64K B for entire message
* 3 octet label
* different 3 octets in the entire message

So roughly 21845.333 elements in the map