@miekg 2019-09-22
Oh I forgot the null byte in my calculation. So it's divided by 4, not 3

On Sat, 21 Sep 2019, 21:43 Charlie Vieth, <notifications@github.com> wrote:

> I don’t believe this is a correct usage of sync.Pool, see golang/go#23199
> <https://github.com/golang/go/issues/23199>.
>
> I think the specific issue that you're referring to is that we are not
> setting an upper bound on the size of objects that we store in the pool.
>
> I'm not terribly familiar with this part of DNS, but maybe @miekg
> <https://github.com/miekg> could weigh in on what an appropriate upper
> limit would be (for reference BenchmarkPackMsgMassive stores 164 elements
> in the map)?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/miekg/dns/pull/1006?email_source=notifications&email_token=AACWIW2Z65CQFQIJ7KV2T6LQK2BQDA5CNFSM4IY4Q2B2YY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOD7IZKNA#issuecomment-533828916>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AACWIW7WALBEYVKSOZSTUDTQK2BQDANCNFSM4IY4Q2BQ>
> .
>