@charlievieth 2019-09-22
> This does look to have real world impact... ?

That just shows that there is tremendous variability (+/- ~50-65%) in the timing of the `Serve` benchmarks.  This existed before this change.