@charlievieth 2019-09-22
So using a max map size of 16,384 (64K B / 4) could lead to storing maps of the following sizes (this is based on how much memory was allocated during map creation):
```
CompressionPackPool (map[string]uint16)    696358 B/op  680.03 KB
CompressionPool     (map[string]struct{})  630816 B/op  616 KB
```
Which isn't terrible, but could lead to a large amount of data residing in the `sync.Pool` in situations where there are many goroutines and the occasional requirement for a large compression map.

This could be handled by bucketing the Pools, but this would be difficult as we don't know how big of a map each request will require.  A compromise might be to use a limit of 8k instead - though this would decrease the effectiveness of the pool when and where it is needed most.

That said, it was CoreDNS that led me here and I don't think this will be an issue in its use case.  The pool always presents the risk of caching large objects, but I think in a steady-state environment this really won't be an issue.

Additionally, re-running the benchmarks on Ubuntu 18.04 (instead of my Mac with its garbage networking) shows an increase in `Serve` performance:
```
benchmark                    old ns/op     new ns/op     delta
BenchmarkServe-4             95307         89116         -6.50%
BenchmarkServe6-4            85613         68566         -19.91%
BenchmarkServeCompress-4     96287         84568         -12.17%
```