@miekg 2019-09-23
Thanks for the benchmarking. Let's merge. The compression pointer `map` is something that I wish I had a better solution for, but hard to do (if you don't want to rewrite the entire lib)