@tmthrgd 2019-09-23
I really do think we'll hit both golang/go#23199 and golang/go#27735 with this change and end up causing memory issues. It's definitely not the correct usage of sync.Pool. I've actually personally moved away from using sync.Pool in many cases because it's quite finicky.