@tmthrgd 2019-09-26
@miekg I think revert. While we may be able to do something here to reduce allocations, this isn’t correct as it stands.