@charlievieth 2019-09-26
@tmthrgd The concerns around `sync.Pool` are real, but in my experience require fairly broken broken code to exercise.  I've been using the `sync.Pool` in large high-throughput systems for years without issue.

It really only becomes and issue when the system is heavily overloaded with goroutines (since this increases the size of the pool - can't steal from `poolLocalInternal.private`) at which point the memory left lingering in the `sync.Pool` is the least of your worries (since the service might not stay up long enough for it to become an issue).

There is a lot of benefit here and its hard to see how CoreDNS or any other responsible consumer of this library will run into the issue.