@miekg 2019-09-26
So the main problem with `sync.Pool` is that we need to bucket the things we put 
in to it. Those need to be roughly of the same size? And this is true even if 
you put empty `map`s in there? Is that in a nutshell what's going wrong here?

The need for a map just to do compression right has long annoyed me (I added it) 
as being very wasteful, so doing something in this area is very much +1