@miekg 2019-09-26
And also having it fail right at the time where we need it most is bad.
Overloading DNS servers is relatively easy so this opens the door to
cascading failure a of all DNS instances. (The current behaviour might be
worse, dunno, but it is more predictable)

On Thu, 26 Sep 2019, 07:17 Miek Gieben, <miek@miek.nl> wrote:

> So the main problem with `sync.Pool` is that we need to bucket the things
> we put
> in to it. Those need to be roughly of the same size? And this is true even
> if
> you put empty `map`s in there? Is that in a nutshell what's going wrong
> here?
>
> The need for a map just to do compression right has long annoyed me (I
> added it)
> as being very wasteful, so doing something in this area is very much +1
>